creating attachments folder.
